fix: Prevent concurrent package publish conflicts
Use the new copyPackageAnonymous() helper to publish anonymous packages
instead of using a shared directory.

Also moves the volume tests to a separate file.
