+++ tests/test.ts
@@ -21,7 +21,6 @@ import {
   randomAppName,
   sleep,
   TestEnv,
-  wasmopticonDir,
   writeAppDefinition,
 } from "../src/index.ts";
 
@@ -144,123 +143,6 @@ Deno.test("app-https-redirect", async () => {
   );
   console.log(await res2.text());
   assertEquals(res2.status, 200);
-});
-
-Deno.test("app-volumes", async () => {
-  const env = TestEnv.fromEnv();
-
-  const phpServerDir = path.join(await wasmopticonDir(), "php/php-testserver");
-  const phpServerPackage = await env.ensurePackagePublished(phpServerDir);
-
-  const app: AppDefinition = {
-    appYaml: {
-      kind: "wasmer.io/App.v0",
-      package: phpServerPackage,
-      // Enable debug mode to allow instance purging.
-      debug: true,
-      volumes: [
-        {
-          name: "data",
-          mount: "/data",
-        },
-      ],
-    },
-  };
-  const info = await env.deployApp(app);
-
-  const file1Content = "value1";
-
-  // Write a file to the volume.
-  await env.fetchApp(info, "/fs/write/data/file1", {
-    method: "POST",
-    body: file1Content,
-    discardBody: true,
-  });
-
-  // Read the file.
-  {
-    const resp = await env.fetchApp(info, "/fs/read/data/file1");
-    const body = await resp.text();
-    assertEquals(body, file1Content);
-  }
-
-  // Now read again, but force a fresh instance to make sure it wasn't just
-  // stored in memory.
-  {
-    const resp = await env.fetchApp(info, "/fs/read/data/file1", {
-      headers: {
-        [HEADER_PURGE_INSTANCES]: "1",
-      },
-    });
-    const body = await resp.text();
-    assertEquals(body, file1Content);
-  }
-});
-
-// Test that a volume can be mounted inside a directory mounted from a package.
-Deno.test("volume-mount-inside-package-dir", async () => {
-  const env = TestEnv.fromEnv();
-
-  const phpServerDir = path.join(await wasmopticonDir(), "php/php-testserver");
-  const phpServerPackage = await env.ensurePackagePublished(phpServerDir);
-
-  // The PHP testserver mounts code at /app, so we'll mount a volume inside that.
-
-  const app: AppDefinition = {
-    appYaml: {
-      kind: "wasmer.io/App.v0",
-      package: phpServerPackage,
-      // Enable debug mode to allow instance purging.
-      debug: true,
-      volumes: [
-        {
-          name: "data",
-          mount: "/app/data",
-        },
-      ],
-    },
-  };
-  const info = await env.deployApp(app);
-
-  const file1Content = "value1";
-
-  const filePath = "/app/data/file1";
-
-  // Write a file to the volume.
-  await env.fetchApp(info, `/fs/write${filePath}`, {
-    method: "POST",
-    body: file1Content,
-    discardBody: true,
-  });
-
-  // Read the file.
-  let firstInstanceId: string;
-  {
-    const resp = await env.fetchApp(info, `/fs/read${filePath}`);
-    const body = await resp.text();
-    assertEquals(body, file1Content);
-    const id = resp.headers.get(HEADER_INSTANCE_ID);
-    assert(id);
-    firstInstanceId = id;
-  }
-  assert(firstInstanceId);
-
-  // Now read again, but force a fresh instance to make sure it wasn't just
-  // stored in memory.
-  {
-    const resp = await env.fetchApp(info, `/fs/read${filePath}`, {
-      headers: {
-        [HEADER_PURGE_INSTANCES]: "1",
-      },
-    });
-    const body = await resp.text();
-    assertEquals(body, file1Content);
-
-    const secondInstanceId = resp.headers.get(HEADER_INSTANCE_ID);
-    assert(secondInstanceId);
-    // Make sure the response was served from a different instance.
-    assertNotEquals(firstInstanceId, secondInstanceId);
-  }
 });
 
 // TODO: fix CGI!
